moving showtime population to db operation file
